requirements.txt
Authlib==1.3.2
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Flask==3.0.3
Flask-Dance==7.1.0
Flask-JWT-Extended==4.6.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
urllib3==2.2.3
URLObject==2.4.3
Werkzeug==3.0.0
